+++ packages/aws-amplify-react/src/Auth/Authenticator.tsx
@@ -179,7 +179,7 @@ export default class Authenticator extends Component<IAuthenticatorProps, IAuthe
         // otherwise if falsey, use EmptyContainer
         const Wrapper = this.props.container === undefined ? Container : this.props.container || EmptyContainer;
 
-        let { hideDefault, hide = [], federated, signUpConfig } = this.props;
+        let { hideDefault, hide = [], federated, signUpConfig, usernameAttributes } = this.props;
         if (hideDefault) {
             hide = hide.concat([
                 Greetings,
@@ -230,7 +230,8 @@ export default class Authenticator extends Component<IAuthenticatorProps, IAuthe
                     onStateChange: this.handleStateChange,
                     onAuthEvent: this.handleAuthEvent,
                     hide,
-                    override: props_children_override
+                    override: props_children_override,
+                    usernameAttributes
                 });
         });
        
@@ -244,7 +245,8 @@ export default class Authenticator extends Component<IAuthenticatorProps, IAuthe
                     onStateChange: this.handleStateChange,
                     onAuthEvent: this.handleAuthEvent,
                     hide,
-                    override: props_children_override
+                    override: props_children_override,
+                    usernameAttributes
                 });
             });