+++ packages/aws-amplify-react/src/Auth/ForgotPassword.tsx
@@ -54,7 +54,7 @@ export default class ForgotPassword extends AuthPiece<IAuthPieceProps, IForgotPa
 
     send() {
         const { authData={} } = this.props;
-        const username = this.inputs.username || authData.username;
+        const username = this.getUsernameFromInput() || authData.username;
         if (!Auth || typeof Auth.forgotPassword !== 'function') {
             throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
         }
@@ -69,7 +69,7 @@ export default class ForgotPassword extends AuthPiece<IAuthPieceProps, IForgotPa
     submit() {
         const { authData={} } = this.props;
         const { code, password } = this.inputs;
-        const username = this.inputs.username || authData.username;
+        const username = this.getUsernameFromInput() || authData.username;
         
         if (!Auth || typeof Auth.forgotPasswordSubmit !== 'function') {
             throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
@@ -87,18 +87,7 @@ export default class ForgotPassword extends AuthPiece<IAuthPieceProps, IForgotPa
         const theme = this.props.theme || AmplifyTheme;
         return (
             <div>
-                <FormField theme={theme}>
-                    <InputLabel theme={theme}>{I18n.get('Username')} *</InputLabel>
-                    <Input
-                        autoFocus
-                        placeholder={I18n.get('Enter your username')}
-                        theme={theme}
-                        key="username"
-                        name="username"
-                        onChange={this.handleInputChange}
-                        data-test={auth.forgotPassword.usernameInput}
-                    />
-                </FormField>
+                {this.renderUsernameField(theme)}
             </div>
         );
     }